requirements.txt
fastapi[all]
langchain-openai
streamlit
requests
python-dotenv